+++ tests/integration/test_xtce_based_parsing/test_jpss_parsing.py
@@ -21,5 +21,7 @@ def test_jpss_xtce_packet_parsing(jpss_test_data_dir):
             assert isinstance(jpss_packet, parser.Packet)
             assert jpss_packet.header['PKT_APID'].raw_value == 11
             assert jpss_packet.header['VERSION'].raw_value == 0
+            assert jpss_packet.data['USEC'].short_description == "Secondary Header Fine Time (microsecond)"
+            assert jpss_packet.data['USEC'].long_description == "CCSDS Packet 2nd Header Fine Time in microseconds."
             n_packets += 1
         assert n_packets == 7200